+++ packages/commonwealth/client/scripts/views/components/community_options_popover.tsx
@@ -9,7 +9,7 @@ import { link } from 'helpers';
 import { ITokenAdapter } from 'models';
 import { NewTopicModal } from 'views/modals/new_topic_modal';
 import { EditTopicThresholdsModal } from 'views/modals/edit_topic_thresholds_modal';
-import CreateInviteModal from 'views/modals/create_invite_modal';
+import { CreateInviteModal } from 'views/modals/create_invite_modal';
 import OrderTopicsModal from '../modals/order_topics_modal';
 import { CWIcon } from './component_kit/cw_icons/cw_icon';
